+++ librisworks/scripts/display-works.groovy
@@ -2,8 +2,7 @@ import se.kb.libris.mergeworks.Doc
 import se.kb.libris.mergeworks.Html
 import se.kb.libris.mergeworks.WorkComparator
 
-import static se.kb.libris.mergeworks.Util.partition
-import static se.kb.libris.mergeworks.Util.sortByIntendedAudience
+import static se.kb.libris.mergeworks.Util.workClusters
 
 htmlReport = getReportWriter('works.html')
 
@@ -19,8 +18,6 @@ new File(System.getProperty('clusters')).splitEachLine(~/[\t ]+/) { cluster ->
                 ?.with { docs.add(new Doc(getWhelk(), it)) }
     }
 
-    sortByIntendedAudience(docs)
-
     WorkComparator c = new WorkComparator(WorkComparator.allFields(docs))
 
     def workClusters = workClusters(docs, c).findAll { it.size() > 1 }
@@ -30,15 +27,6 @@ new File(System.getProperty('clusters')).splitEachLine(~/[\t ]+/) { cluster ->
 }
 
 htmlReport.println(Html.END)
-
-Collection<Collection<Doc>> workClusters(Collection<Doc> docs, WorkComparator c) {
-    docs.each { it.addComparisonProps() }
-
-    def workClusters = partition(docs, { Doc a, Doc b -> c.sameWork(a, b) })
-            .each { work -> work.each { doc -> doc.removeComparisonProps() } }
-
-    return workClusters
-}
 
 Doc createNewWork(Map workData) {
     workData['@id'] = "TEMPID#it"
